+++ src/model.jl
@@ -8,22 +8,32 @@ Create the model and call all required functions.
   it is the case for standard `EnergyModelsBase` models. In addition, the keys `:areas` and
   `:transmission` are required for extending the existing model.
   If the input is not provided in the correct form, the checks will identify the problem.
-  In the case of a
 - `modeltype::EnergyModel` - Used modeltype, that is a subtype of the type `EnergyModel`.
 - `m` - the empty `JuMP.Model` instance. If it is not provided, then it is assumed that the
   input is a standard `JuMP.Model`.
 
 ## Conditional input
-- `check_timeprofiles::Bool=true` - A boolean indicator whether the time profiles of the individual
-  nodes should be checked or not. It is advised to not deactivate the check, except if you
-  are testing new components. It may lead to unexpected behaviour and potential
-  inconsistencies in the input data, if the time profiles are not checked.
+- `check_timeprofiles::Bool=true` - A boolean indicator whether the time profiles of the
+  individual nodes should be checked or not. It is advised to not deactivate the check,
+  except if you are testing new components. It may lead to unexpected behaviour and
+  potential inconsistencies in the input data, if the time profiles are not checked.
+- `check_any_data::Bool=true` - A boolean indicator whether the input data is checked or not.
+  It is advised to not deactivate the check, except if you are testing new features.
+  It may lead to unexpected behaviour and even infeasible models.
 """
-function create_model(case, modeltype::EnergyModel, m::JuMP.Model; check_timeprofiles::Bool=true)
+function create_model(
+    case,
+    modeltype::EnergyModel,
+    m::JuMP.Model;
+    check_timeprofiles::Bool=true,
+    check_any_data::Bool = true,
+)
     @debug "Construct model"
     # Call of the basic model
     m = EMB.create_model(case, modeltype, m; check_timeprofiles)
-    check_data(case, modeltype, check_timeprofiles)
+    if check_any_data
+        check_data(case, modeltype, check_timeprofiles)
+    end
 
     # Data structure
     𝒜 = case[:areas]
@@ -54,16 +64,22 @@ function create_model(case, modeltype::EnergyModel, m::JuMP.Model; check_timepro
 
     return m
 end
-function create_model(case, modeltype::EnergyModel; check_timeprofiles::Bool=true)
+function create_model(
+    case,
+    modeltype::EnergyModel;
+    check_timeprofiles::Bool = true,
+    check_any_data::Bool = true,
+)
     m = JuMP.Model()
-    create_model(case, modeltype, m; check_timeprofiles)
+    create_model(case, modeltype, m; check_timeprofiles, check_any_data)
 end
 
 """
     variables_area(m, 𝒜, 𝒯, ℒᵗʳᵃⁿˢ, modeltype::EnergyModel)
 
-Create variables to track how much energy is exchanged from an area for all
-time periods `t ∈ 𝒯`.
+Declaration of a variable `:area_exchange` to track how much energy is exchanged from an
+area for all operational periods `t ∈ 𝒯`. The variable is only declared for resources that
+are exchanged from a given area.
 """
 function variables_area(m, 𝒜, 𝒯, ℒᵗʳᵃⁿˢ, modeltype::EnergyModel)
     @variable(m, area_exchange[a ∈ 𝒜, 𝒯, p ∈ exchange_resources(ℒᵗʳᵃⁿˢ, a)])
@@ -80,10 +96,10 @@ function variables_trans_capex(m, 𝒯, ℳ, modeltype::EnergyModel) end
 """
     variables_trans_opex(m, 𝒯, ℳ, modeltype::EnergyModel)
 
-Declaration of variables for the operational costs of the tranmission modes.
+Declaration of variables for the operational costs (`:trans_opex_var` and
+`:trans_opex_fixed`) of the tranmission modes.
 """
 function variables_trans_opex(m, 𝒯, ℳ, modeltype::EnergyModel)
-
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
 
     @variable(m, trans_opex_var[ℳ, 𝒯ᴵⁿᵛ])
@@ -93,8 +109,8 @@ end
 """
     variables_trans_capacity(m, 𝒯, ℳ, modeltype::EnergyModel)
 
-Declaration of variables for tracking how much of installed transmision capacity is used f
-or all time periods `t ∈ 𝒯` of the tranmission modes.
+Declaration of variables for tracking how much of installed transmision capacity is used in
+all operational periods `t ∈ 𝒯` of the tranmission modes.
 """
 function variables_trans_capacity(m, 𝒯, ℳ, modeltype::EnergyModel)
 
@@ -108,7 +124,8 @@ Loop through all `TransmissionMode` types and create variables specific to each 
 This is done by calling the method [`variables_trans_mode`](@ref) on all modes of each type.
 
 The `TransmissionMode` type representing the widest category will be called first. That is,
-`variables_trans_mode` will be called on a `TransmissionMode` before it is called on `PipeMode`-nodes.
+`variables_trans_mode` will be called on a `TransmissionMode` before it is called on
+`PipeMode`.
 """
 function variables_trans_modes(m, 𝒯, ℳ, modeltype::EnergyModel)
 
@@ -139,10 +156,10 @@ end
 """
     variables_trans_mode(m, 𝒯, ℳˢᵘᵇ::Vector{<:TransmissionMode}, modeltype::EnergyModel)
 
-Default fallback method when no function is defined for a `TransmissionMode` type.
-It introduces the variables that are required in all `TransmissionMode`s. These variables
-are:
+Default fallback method when no function is defined for a [`TransmissionMode`](@ref) type.
+It introduces the variables that are required in all `TransmissionMode`s.
 
+These variables are:
 * `:trans_in` - inlet flow to transmission mode
 * `:trans_out` - outlet flow from a transmission mode
 * `:trans_loss` - loss during transmission
@@ -165,9 +182,8 @@ end
 """
     variables_trans_mode(m, 𝒯, ℳᴸᴾ::Vector{<:PipeLinepackSimple}, modeltype::EnergyModel)
 
-Adds the following special variables for linepacking:
-
-* `:linepack_stor_level` - storage level in linepack
+When the node vector is a `Vector{<:PipeLinepackSimple}`, we declare the variable
+`:linepack_stor_level` to account for the energy stored through line packing.
 """
 function variables_trans_mode(m, 𝒯, ℳᴸᴾ::Vector{<:PipeLinepackSimple}, modeltype::EnergyModel)
 
@@ -177,12 +193,12 @@ end
 """
     variables_trans_emission(m, 𝒯, ℳ, 𝒫, modeltype)
 
-Creates variables for the modeling of tranmission emissions. These variables
-are only created for transmission modes where emissions are included.
-All emission resources that are not included for a type are fixed to a value of 0.
+Declation of variables for the modeling of tranmission emissions. These variables are only
+created for transmission modes where emissions are included. All emission resources that are
+not included for a type are fixed to a value of 0.
 
 The emission variables are differentiated in:
-* `:emissions_node` - emissions of a transmission mode in an operational period,
+* `:emissions_trans` - emissions of a transmission mode `tm` in operational period `t`.
 """
 function variables_trans_emission(m, 𝒯, ℳ, 𝒫, modeltype)
     ℳᵉᵐ = filter(m -> has_emissions(m), ℳ)
@@ -199,8 +215,10 @@ end
 """
     constraints_area(m, 𝒜, 𝒯, ℒᵗʳᵃⁿˢ, 𝒫, modeltype::EnergyModel)
 
-Create constraints for the energy balances within an area for each resource using the GeoAvailability node.
-Keep track of the exchange with other areas in a seperate variable `:area_exchange`.
+Function for creating constraints for the energy balances within an area for each resource
+using the [`GeoAvailability`](@ref) node. It keeps track of the exchange with other areas
+through the variable `:area_exchange` and the functions [`compute_trans_in`](@ref) and
+[`compute_trans_out`](@ref).
 """
 function constraints_area(m, 𝒜, 𝒯, ℒᵗʳᵃⁿˢ, 𝒫, modeltype::EnergyModel)
     for a ∈ 𝒜
@@ -236,22 +254,24 @@ end
 """
     EMB.create_node(m, n::GeoAvailability, 𝒯, 𝒫, modeltype::EnergyModel)
 
-Repaces constraints for availability nodes of type GeoAvailability.
-The resource balances are set by the area constraints instead.
+Set all constraints for a [`GeoAvailability`](@ref). The energy balance is handlded in the
+function [`constraints_area`](@ref). Hence, no constraints are added
 """
 function EMB.create_node(m, n::GeoAvailability, 𝒯, 𝒫, modeltype::EnergyModel) end
 
 """
     create_area(m, a::Area, 𝒯, ℒᵗʳᵃⁿˢ, modeltype)
 
-Default fallback method when no function is defined for a node type.
+Set all constraints for an [`Area`](@ref). Can serve as fallback option for all unspecified
+subtypes of `Area`.
 """
 function create_area(m, a::Area, 𝒯, ℒᵗʳᵃⁿˢ, modeltype) end
 
 """
     create_area(m, a::LimitedExchangeArea, 𝒯, ℒᵗʳᵃⁿˢ, modeltype)
 
-Constraint that limit exchange with other areas based on the specified exchange_limit.
+When the area is a [`LimitedExchangeArea`](@ref), we limit the export of the specified
+limit resources `p` to the providewd value.
 """
 function create_area(m, a::LimitedExchangeArea, 𝒯, ℒᵗʳᵃⁿˢ, modeltype)
 
@@ -277,7 +297,7 @@ function constraints_transmission(m, 𝒯, ℳ, modeltype::EnergyModel)
 end
 
 """
-    update_objective(m, 𝒩, 𝒯, 𝒫, ℒᵗʳᵃⁿˢ, modeltype::EnergyModel)
+    update_objective(m, 𝒯, ℳ, modeltype::EnergyModel)
 
 Update the objective function with costs related to geography (areas and energy transmission).
 """
@@ -287,20 +307,21 @@ function update_objective(m, 𝒯, ℳ, modeltype::EnergyModel)
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
     obj = objective_function(m)
 
-    # Update of the cost function for modes with investments
-    for t_inv ∈ 𝒯ᴵⁿᵛ, tm ∈ ℳ
-        obj -= duration_strat(t_inv) * m[:trans_opex_fixed][tm, t_inv]
-        obj -= duration_strat(t_inv) * m[:trans_opex_var][tm, t_inv]
-    end
-
-    @objective(m, Max, obj)
+    # Calculate the OPEX cost contribution
+    opex = @expression(m, [t_inv ∈ 𝒯ᴵⁿᵛ],
+        sum(m[:trans_opex_var][tm, t_inv] + m[:trans_opex_fixed][tm, t_inv] for tm ∈ ℳ)
+    )
+    # Update the objective
+    @objective(m, Max,
+        obj - sum(opex[t_inv] * duration_strat(t_inv) for t_inv ∈ 𝒯ᴵⁿᵛ)
+    )
 end
 
 """
     update_total_emissions(m, 𝒯, ℳ, 𝒫, modeltype::EnergyModel)
 
-Update the constraints aggregating total emissions in each time period
-with contributions from transmission emissions.
+Update the constraints aggregating total emissions in each operational period with
+contributions from transmission emissions.
 """
 function update_total_emissions(m, 𝒯, ℳ, 𝒫, modeltype::EnergyModel)
     ℳᵉᵐ = filter(m -> has_emissions(m), ℳ)
@@ -310,15 +331,28 @@ function update_total_emissions(m, 𝒯, ℳ, 𝒫, modeltype::EnergyModel)
     # transmission emissions. Note the coefficient is set to -1 since the total constraint
     # has the variables on the RHS.
     for tm ∈ ℳᵉᵐ, p ∈ 𝒫ᵉᵐ, t ∈ 𝒯
-        JuMP.set_normalized_coefficient(m[:con_em_tot][t, p], m[:emissions_trans][tm, t, p], -1.0)
+        JuMP.set_normalized_coefficient(
+            m[:con_em_tot][t, p],
+            m[:emissions_trans][tm, t, p],
+            -1.0,
+        )
     end
 end
 
 """
     create_transmission_mode(m, tm::TransmissionMode, 𝒯, modeltype::EnergyModel)
 
-Set all constraints for transmission mode. Serves as a fallback option for unspecified
-subtypes of `TransmissionMode`.
+Set all constraints for a [`TransmissionMode`](@ref).
+
+Serves as a fallback option for unspecified subtypes of `TransmissionMode`.
+
+# Called constraint functions
+- [`constraints_trans_balance`](@ref),
+- [`constraints_trans_loss`](@ref),
+- [`constraints_capacity`](@ref),
+- [`constraints_emission`](@ref),
+- [`constraints_opex_fixed`](@ref), and
+- [`constraints_opex_var`](@ref).
 """
 function create_transmission_mode(m, tm::TransmissionMode, 𝒯, modeltype::EnergyModel)