+++ src/structures/mode.jl
@@ -1,16 +1,20 @@
-""" Declaration of the general type for transmission mode transporting resources between areas."""
+"""
+    abstract type TransmissionMode
+
+Declaration of the general type for transmission modes transporting resources between areas.
+"""
 abstract type TransmissionMode end
 Base.show(io::IO, t::TransmissionMode) = print(io, "$(t.id)")
 
 """
-    RefDynamic <: TransmissionMode
+    struct RefDynamic <: TransmissionMode
 
 A reference dynamic `TransmissionMode`.
 
 Generic representation of dynamic transmission modes, using for example truck, ship or railway transport.
 
 # Fields
-- **`id::String`** is the name/identifier of the transmission mode.
+- **`id::String`** is the name/identifyer of the transmission mode.
 - **`resource::Resource`** is the resource that is transported.
 - **`trans_cap::TimeProfile`** is the capacity of the transmission mode.
 - **`trans_loss::TimeProfile`** is the loss of the transported resource during
@@ -19,10 +23,10 @@ Generic representation of dynamic transmission modes, using for example truck, s
 - **`opex_fixed::TimeProfile`** is the fixed operating expense per installed capacity.
 - **`directions`** is the number of directions the resource can be transported,
   1 is unidirectional (A->B) or 2 is bidirectional (A<->B).
-- **`data::Vector{Data}`** is the additional data (e.g. for investments). The field `data`
+- **`data::Vector{Data}`** is the additional data (*e.g.*, for investments). The field `data`
   is conditional through usage of a constructor.
 """
-struct RefDynamic <: TransmissionMode # E.g. Trucks, ships etc.
+struct RefDynamic <: TransmissionMode # *e.g.*, Trucks, ships etc.
     id::String
     resource::EMB.Resource
     trans_cap::TimeProfile
@@ -45,14 +49,14 @@ function RefDynamic(
 end
 
 """
-    RefStatic <: TransmissionMode
+    struct RefStatic <: TransmissionMode
 
 A reference static `TransmissionMode`.
 
 Generic representation of static transmission modes, such as overhead power lines or pipelines.
 
 # Fields
-- **`id::String`** is the name/identifier of the transmission mode.
+- **`id::String`** is the name/identifyer of the transmission mode.
 - **`resource::Resource`** is the resource that is transported.
 - **`trans_cap::Real`** is the capacity of the transmission mode.
 - **`trans_loss::Real`** is the loss of the transported resource during transmission,
@@ -61,7 +65,7 @@ Generic representation of static transmission modes, such as overhead power line
 - **`opex_fixed::TimeProfile`** is the fixed operating expense per installed capacity.
 - **`directions`** is the number of directions the resource can be transported,
   1 is unidirectional (A->B) or 2 is bidirectional (A<->B).
-- **`data::Vector{Data}`** is the additional data (e.g. for investments). The field `data`
+- **`data::Vector{Data}`** is the additional data (*e.g.*, for investments). The field `data`
   is conditional through usage of a constructor.
 """
 struct RefStatic <: TransmissionMode # E.g. overhead power lines, pipelines etc.
@@ -88,18 +92,18 @@ end
 
 
 """
-    PipeMode <: TransmissionMode
+    abstract type PipeMode <: TransmissionMode
 
 `TransmissionMode` mode for additional variable potential.
 """
 abstract type PipeMode <: TransmissionMode end
 
 """
-    PipeSimple <: PipeMode
+    struct PipeSimple <: PipeMode
 
 This `TransmissionMode` allows for altering the transported `Resource`.
 
-A usage of this could e.g. be by defining a subtype struct of `Resource` with the field
+A usage of this could be, *e.g.*, by defining a subtype struct of `Resource` with the field
 'pressure'. This PipelineMode can then take `SomeSubtype<:Resource` with pressure p₁ at the
 inlet, and pressure p₂ at the outlet.
 
@@ -116,15 +120,16 @@ be consumed at the wrong `Area`.
 - **`inlet::Resource`** is the `Resource` going into transmission.
 - **`outlet::Resource`** is the `Resource` going out of the outlet of the transmission.
 - **`consuming::Resource`** is the `Resource` the transmission consumes by operating.
-- **`consumption_rate::Real`** the rate of which the resource `Pipeline.consuming` is
-  consumed, as a ratio of the volume of the resource going into the inlet, i.e.:\n
+- **`consumption_rate::TimeProfile`** the rate of which the resource `Pipeline.consuming` is
+  consumed, as a ratio of the volume of the resource going into the inlet, *i.e.*:
+
         `consumption_rate` = consumed volume / inlet volume (per operational period)
 - **`trans_cap::Real`** is the capacity of the transmission mode.
 - **`trans_loss::Real`** is the loss of the transported resource during transmission,
   modelled as a ratio.
 - **`opex_var::TimeProfile`** is the variable operating expense per energy unit transported.
 - **`opex_fixed::TimeProfile`** is the fixed operating expense per installed capacity.
-- **`data::Array{<:Data}`** is the additional data (e.g. for investments). The field `data`
+- **`data::Vector{Data}`** is the additional data (*e.g.*, for investments). The field `data`
   is conditional through usage of a constructor.
 """
 struct PipeSimple <: PipeMode
@@ -193,17 +198,17 @@ function PipeSimple(
 end
 
 """
-    PipeLinepackSimple <: PipeMode
+    struct PipeLinepackSimple <: PipeMode
 
 Pipeline model with linepacking implemented as simple storage function.
 
-# Fields (additional to `PipeSimple`)
+# Fields
 - **`id::String`** is the identifier used in printed output.
 - **`inlet::Resource`** is the `Resource` going into transmission.
 - **`outlet::Resource`** is the `Resource` going out of the outlet of the transmission.
 - **`consuming::Resource`** is the `Resource` the transmission consumes by operating.
-- **`consumption_rate::Real`** the rate of which the resource `Pipeline.consuming` is
-  consumed, as a ratio of the volume of the resource going into the inlet, i.e.:\n
+- **`consumption_rate::TimeProfile`** the rate of which the resource `Pipeline.consuming` is
+  consumed, as a ratio of the volume of the resource going into the inlet, *i.e.*:\n
         `consumption_rate` = consumed volume / inlet volume (per operational period)
 - **`trans_cap::Real`** is the capacity of the transmission mode.
 - **`trans_loss::Real`** is the loss of the transported resource during transmission,
@@ -211,7 +216,7 @@ Pipeline model with linepacking implemented as simple storage function.
 - **`opex_var::TimeProfile`** is the variable operating expense per energy unit transported.
 - **`opex_fixed::TimeProfile`** is the fixed operating expense per installed capacity.
 - **`energy_share::Float64`** is the storage energy capacity relative to pipeline capacity.
-- **`data::Array{<:Data}`** is the additional data (e.g. for investments). The field `data`
+- **`data::Array{<:Data}`** is the additional data (*e.g.*, for investments). The field `data`
   is conditional through usage of a constructor.
 """
 struct PipeLinepackSimple <: PipeMode
@@ -319,7 +324,8 @@ function modes_sub(ℳ::Vector{<:TransmissionMode}, string_array::Array{String})
 end
 
 """
-    map_trans_resource(tm)
+    map_trans_resource(tm::TransmissionMode)
+    map_trans_resource(tm::PipeMode)
 
 Returns the transported resource for a given TransmissionMode.
 """
@@ -338,19 +344,15 @@ end
 
 """
     capacity(tm::TransmissionMode)
-
-Returns the capacity of transmission mode `tm` as `TimeProfile`.
-"""
-EMB.capacity(tm::TransmissionMode) = tm.trans_cap
-"""
     capacity(tm::TransmissionMode, t)
 
-Returns the capacity of transmission mode `tm` at time period `t`.
+Returns the capacity of transmission mode `tm` as `TimeProfile` or in operational period `t`.
 """
+EMB.capacity(tm::TransmissionMode) = tm.trans_cap
 EMB.capacity(tm::TransmissionMode, t) = tm.trans_cap[t]
-
 """
-    input(tm::TransmissionMode)
+    inputs(tm::TransmissionMode)
+    inputs(tm::PipeMode)
 
 Returns the input resources of transmission mode `tm`.
 """
@@ -358,7 +360,8 @@ EMB.inputs(tm::TransmissionMode) = [tm.resource]
 EMB.inputs(tm::PipeMode) = [tm.inlet, tm.consuming]
 
 """
-    output(tm::TransmissionMode)
+    outputs(tm::TransmissionMode)
+    outputs(tm::PipeMode)
 
 Returns the output resources of transmission mode `tm`.
 """
@@ -367,42 +370,31 @@ EMB.outputs(tm::PipeMode) = [tm.outlet]
 
 """
     opex_var(tm::TransmissionMode)
-
-Returns the variable OPEX of transmission mode `tm` as `TimeProfile`.
-"""
-EMB.opex_var(tm::TransmissionMode) = tm.opex_var
-
-"""
     opex_var(tm::TransmissionMode, t)
 
-Returns the variable OPEX of transmission mode `tm` at time period `t`.
+Returns the variable OPEX of transmission mode `tm` as `TimeProfile` or in operational
+period `t`.
 """
+EMB.opex_var(tm::TransmissionMode) = tm.opex_var
 EMB.opex_var(tm::TransmissionMode, t) = tm.opex_var[t]
 
 """
     opex_fixed(tm::TransmissionMode)
-
-Returns the variable OPEX of transmission mode `tm` as `TimeProfile`.
-"""
-EMB.opex_fixed(tm::TransmissionMode) = tm.opex_fixed
-"""
     opex_fixed(tm::TransmissionMode, t_inv)
 
-Returns the variable OPEX of transmission mode `tm` at strategic period `t_inv`.
+Returns the variable OPEX of transmission mode `tm` as `TimeProfile` or in strategic period
+`t_inv`.
 """
+EMB.opex_fixed(tm::TransmissionMode) = tm.opex_fixed
 EMB.opex_fixed(tm::TransmissionMode, t_inv) = tm.opex_fixed[t_inv]
 
 """
     loss(tm::TransmissionMode)
-
-Returns the loss of transmission mode `tm` as `TimeProfile`.
-"""
-loss(tm::TransmissionMode) = tm.trans_loss
-"""
     loss(tm::TransmissionMode, t)
 
-Returns the loss of transmission mode `tm` at time period `t`.
+Returns the loss of transmission mode `tm` as `TimeProfile` or in operational period `t`.
 """
+loss(tm::TransmissionMode) = tm.trans_loss
 loss(tm::TransmissionMode, t) = tm.trans_loss[t]
 
 """
@@ -429,30 +421,26 @@ emit_resources(tm::TransmissionMode) = EMB.ResourceEmit[]
 
 """
     emission(tm::TransmissionMode, p::EMB.ResourceEmit)
-
-Returns the emission of transmission mode `tm` of a specific resource `p` as `TimeProfile`
-"""
-emission(tm::TransmissionMode, p::EMB.ResourceEmit) = 0
-
-"""
     emission(tm::TransmissionMode, p::EMB.ResourceEmit, t)
 
-Returns the emission of transmission mode `tm` of a specific resource `p` at time period `t`
-per unit transmitted.
+Returns the emission of transmission mode `tm` of a specific resource `p` as `TimeProfile`
+or in operational period `ts`.
+
+!!! note "Transmission emissions"
+    None of the provided `TransmissionMode`s include emissions. If you plan to include
+    emissions, you have to both create a new `TransmissionMode` and dispatch on this
+    function.
 """
+emission(tm::TransmissionMode, p::EMB.ResourceEmit) = 0
 emission(tm::TransmissionMode, p::EMB.ResourceEmit, t) = 0
 
 """
     consumption_rate(tm::PipeMode)
-
-Returns the consumption rate of pipe mode `tm` as `TimeProfile`.
-"""
-consumption_rate(tm::PipeMode) = tm.consumption_rate
-"""
     consumption_rate(tm::PipeMode, t)
 
-Returns the consumption rate of pipe mode `tm` at time period `t`.
+Returns the consumption rate of pipe mode `tm` as `TimeProfile` or in operational period `t`.
 """
+consumption_rate(tm::PipeMode) = tm.consumption_rate
 consumption_rate(tm::PipeMode, t) = tm.consumption_rate[t]
 
 """
@@ -465,6 +453,13 @@ energy_share(tm::PipeLinepackSimple) = tm.energy_share
 """
     is_bidirectional(tm::TransmissionMode)
 
-Checks whether TransmissionMode `tm` is bidirectional.
+Checks whether transmission mode `tm` is bidirectional.
 """
 is_bidirectional(tm::TransmissionMode) = tm.directions == 2
+
+"""
+    mode_data(tm::TransmissionMode)
+
+Returns the [`Data`](@extref EnergyModelsBase.Data) array of transmission mode `tm`.
+"""
+mode_data(tm::TransmissionMode) = tm.data