+++ test/runtests.jl
@@ -5,24 +5,44 @@ using Test
 using EnergyModelsBase
 using TimeStruct
 
-
 const TS = TimeStruct
 const EMB = EnergyModelsBase
 const EMG = EnergyModelsGeography
 
-const TEST_ATOL = 1e-6
-const ROUND_DIGITS = 8
-
+include("utils.jl")
 
 @testset "Geography" begin
-    include("utils.jl")
 
-    include("test_examples.jl")
-    include("test_geo_unidirectional.jl")
-    include("test_geo_bidirectional.jl")
-    include("test_geo_opex.jl")
-    include("test_simplepipe.jl")
-    include("test_simplelinepack.jl")
-    include("test_utils.jl")
-    include("test_checks.jl")
+    @testset "Geography | Reference modes" begin
+        include("test_geo_unidirectional.jl")
+        include("test_geo_bidirectional.jl")
+    end
+
+    @testset "Geography | OPEX" begin
+        include("test_geo_opex.jl")
+    end
+
+    @testset "Geography | SimplePipe" begin
+        include("test_simplepipe.jl")
+    end
+
+    @testset "Geography | SimpleLinePack" begin
+        include("test_simplelinepack.jl")
+    end
+
+    @testset "Geography | Utilities" begin
+        include("test_utils.jl")
+    end
+
+    @testset "Geography | Checks" begin
+        include("test_checks.jl")
+    end
+
+    @testset "Geography | Investments" begin
+        include("test_investments.jl")
+    end
+
+    @testset "Geography | examples" begin
+        include("test_examples.jl")
+    end
 end